Add ai integgorator, rewrite some code.

+++ src/main/java/com/gallery/galleryapplication/models/Interfaces/ImageProvider.java
@@ -8,5 +8,5 @@ public interface ImageProvider {
     public int getMediaId();
     public String getTagsInString();
     public List<Tag> getTags();
-    public double[] getEmbedding();
+    public float[] getEmbedding();
 }
